// src/api/api.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/globalInfo": {
    /** Get global info */
    get: operations["getGlobalInfo"];
  };
  "/assets": {
    /** Get a list of supported assets */
    get: operations["getAssets"];
  };
}

export interface components {
  schemas: {
    GlobalInfo: {
      nextReallocationTime?: string;
      contractAddresses?: {
        // cBridge: string;
        rollupChain: string;
        faucet?: string;
      };
    };
    Assets: {
      assets: components["schemas"]["Asset"][];
    };
    Asset: {
      id: number;
      address: string;
      name: string;
      symbol: string;
      decimal: number;
      icon: string;
      checked?: boolean;
      title: string;
    };
  };
}

export interface operations {
  /** Get global info */
  getGlobalInfo: {
    responses: {
      /** successful operation */
      200: {
        content: {
          "application/json": components["schemas"]["GlobalInfo"];
        };
      };
    };
  };
  /** Get a list of supported assets */
  getAssets: {
    parameters: {
      query: {
        /** User eth address */
        user?: string;
      };
    };
    responses: {
      /** successful operation */
      200: {
        content: {
          "application/json": components["schemas"]["Assets"];
        };
      };
    };
  };
}